flytt filter-state for oppgaver til egen fil
