+++ frontend/mr-admin-flate/src/api/oppgaver/useOppgaver.ts
@@ -1,7 +1,7 @@
 import { QueryKeys } from "@/api/QueryKeys";
-import { OppgaverService, Tiltakskode } from "@mr/api-client-v2";
+import { OppgaverService } from "@mr/api-client-v2";
 import { useApiQuery } from "@mr/frontend-common";
-import { OppgaverFilterType } from "../atoms";
+import { OppgaverFilterType } from "@/pages/oppgaveoversikt/oppgaver/filter";
 
 export function useOppgaver(filter: OppgaverFilterType) {
   return useApiQuery({
@@ -9,7 +9,7 @@ export function useOppgaver(filter: OppgaverFilterType) {
     queryFn: () =>
       OppgaverService.getOppgaver({
         body: {
-          tiltakskoder: filter.tiltakstyper as Tiltakskode[],
+          tiltakskoder: filter.tiltakstyper,
           oppgavetyper: filter.type,
           regioner: filter.regioner,
         },
