+++ frontend/mr-admin-flate/src/pages/oppgaveoversikt/oppgaver/filter.ts
@@ -1,0 +1,27 @@
+import { atom } from "jotai";
+import { OppgaveType, Tiltakskode } from "@mr/api-client-v2";
+import { z } from "zod";
+import { createFilterStateAtom } from "@/filter/filter-state";
+import { createFilterValidator } from "@/api/atoms";
+
+const OppgaverFilterSchema = z.object({
+  type: z.nativeEnum(OppgaveType).array(),
+  tiltakstyper: z.nativeEnum(Tiltakskode).array(),
+  regioner: z.array(z.string()),
+});
+
+export type OppgaverFilterType = z.infer<typeof OppgaverFilterSchema>;
+
+const defaultOppgaverFilter: OppgaverFilterType = {
+  type: [],
+  tiltakstyper: [],
+  regioner: [],
+};
+
+export const oppgaverFilterStateAtom = createFilterStateAtom<OppgaverFilterType>(
+  "oppgaver-filter",
+  defaultOppgaverFilter,
+  createFilterValidator(OppgaverFilterSchema),
+);
+
+export const oppgaverFilterAccordionAtom = atom<string[]>(["type", "regioner"]);
